React

Basic application

This module is laid out as both a tutorial and documentation. You should be able to read it like a tutorial from start to end.

The way to start a React application within an element is via the react function:

react

Arguments

:: (Show state, Eq state)

=> App state m
The application's state.

-> (state -> ReactT state m ())
The view renderer.

-> Element
The parent DOM element into which the application is placed.

-> IO ()

Loops forever. Blocks on state updates; whenever state changes it re-renders.

The app type

The App type is an opaque type which holds a mutable cell containing the pure state value.

data App state m

An application with some state transforming over some monad.

To make an App value, you can use makeApp:

makeApp

Arguments

:: Monad m

=> state
The initial application state.

-> (forall a. m a -> IO a)
The monad runner. By default you can use id.

-> IO (App state m)
A new application state.

Make the application with the given initial state and runner.

Render function

The react function also needs a rendering function. This should receive the current state and run in the ReactT monad.

For example:

\state -> build "div" (text "Hello, World!")

This will create a <div>Hello, World!</div> element.

You can find a complete list of HTML builder combinators in React.Builder.

data ReactT state m a

React transformer.

The parent DOM element

React needs a parent element. You can use this handy getElementById below to find it.

data Element

A DOM element.

getElementById

Arguments

:: JSString
Element identifier.

-> IO Element

Get the element by its id="foo" identifer..

Building the DOM

There are several basic building functions that you can use to render your reactive DOM tree.

build

Arguments

:: Monad m

=> Text
Name of the element.

-> ReactT state m a
Inner content.

-> ReactT state m a

Build an element.

text

Arguments

:: Monad m

=> Text
Text content.

-> ReactT state m ()

Add some text to the current node's children.

style

Arguments

:: Monad m

=> Text
CSS property name in JavaScript format; fontSize, etc.

-> Text
Value.

-> ReactT state m ()

Add a style.

attr

Arguments

:: Monad m

=> Text
Name.

-> Text
Value.

-> ReactT state m ()

Add attributes. Does not overwrite existing keys.

styles :: Monad m => [(Text, Text)] -> ReactT state m ()

Add styles. Does not overwrite existing keys.

attrs :: Monad m => [(Text, Text)] -> ReactT state m ()

Add attributes. Does not overwrite existing keys.

Build components

A component is an element which manages its own state, but that state is managed by a Lens into the main application state.

So when building a component with buildComponent, we pass in a lens which lens that component access the state.

buildComponent

Arguments

:: Monad m

=> Component s a m
The component.

-> Lens s s a a
A cursor into the state for this instance.

-> ReactT s m x
Set attributes for the component. Ignores content (for now).

-> ReactT s m x

Build a component.

newClass

Arguments

:: App state m
Application.

-> ReactT state m ()
Rendering function.

-> (forall props. Lens' state cursor -> JQuery -> JSRef props -> IO ())
Did mount handler.

-> (forall props. Lens' state cursor -> JSRef props -> IO ())
Did update.

-> (forall props. Lens' state cursor -> JSRef props -> IO Bool)
Should update?

-> (forall props. Lens' state cursor -> JSRef props -> IO ())
Receiving new props.

-> Class state cursor m

Make a new class spec.

createComponent :: Monad m => Class state cursor m -> IO (Component state cursor m)

Create a component class.

data Class state cursor m

Class used for creating components.

classApp :: Class state cursor m -> App state m

Get the app of the class.

Handling events

There is one main function for binding event handlers which is onEvent (see below). There are also a few basic helper functions like onClick, onMouseOut, etc.

For example:

build "button" (do text "Click me!"
                   onClick (event state -> putStrLn "You clicked me!"))

There is a variety of event predicates and accessors which can be found in React.Events.

onEvent :: (IsReactEvent event, Monad m) => EventListener event -> (event -> TVar state -> IO ()) -> ReactT state m ()

onClick :: Monad m => (MouseEvent -> TVar state -> IO ()) -> ReactT state m ()

onDblClick :: Monad m => (MouseEvent -> TVar state -> IO ()) -> ReactT state m ()

onMouseMove :: Monad m => (MouseEvent -> TVar state -> IO ()) -> ReactT state m ()

onMouseOut :: Monad m => (MouseEvent -> TVar state -> IO ()) -> ReactT state m ()