React.Internal

Description

An Om-inspired API for React.js from Haskell.

Documentation

data App state m

An application with some state transforming over some monad.

Constructors

App

Fields

appState :: TVar state

State used for rendering.

appRun :: forall a. m a -> IO a

Function for running in the particular monad.

appIdSource :: TVar Int

Source of integers.

appCursors :: TVar (HashMap Int Cursor)

Mapping from integer to state cursors.

data Cursor

State cursor.

Constructors

forall cursor state . Cursor (Lens cursor state)

data ReactElement_

A virtual react element.

type ReactElement' = JSRef ReactElement_

data ReactClass_

A React component class.

type ReactClass' a = JSRef ReactClass_

data ReactEvent_

Not sure what's in here. Perhaps click coordinates and things like that.

type ReactEvent = JSRef ReactEvent_

data ReactNode_

type ReactNode' = JSRef ReactNode_

toPropsRef :: App state m -> ElemProps -> Maybe (Lens state cursor) -> IO (JSRef props)

Create a JS object for the properties.

Ensures it creates a "key" element for React to use.

Style object is converted to a straight key->string object.

For each of the event handlers it creates a foreign JS wrapper and assigns it in the object.

Example:

{"key": 123123123, onClick: function, style: {textDecoration: "underline"}}

js_React_render :: ReactElement' -> Element -> IO ()

First argument is the virtual element, latter is a real DOM element serving as the container.

js_React_props_cursor :: JSRef this -> IO Int

Create a virtual ReactElement which can be rendered efficiently.

js_React_createElementFromClass :: ReactClass' state -> JSRef props -> JSRef (Lens state cursor) -> IO ReactElement'

dom_a :: IO ReactElement'

js_React_createElement :: JSString -> JSRef props -> ReactNode' -> IO ReactElement'

js_React_createClass :: (JSObject ReactElement' -> IO ()) -> (JSRef JQuery -> JSRef props -> IO ()) -> (JSRef props -> IO ()) -> (JSRef props -> IO ()) -> (JSRef props -> JSRef props -> IO ()) -> IO (ReactClass' state)

class ToReactNode state a where

Methods

toReactNode :: App state m -> a -> IO ReactNode'

data ElemProps

Separated because different types of properties will have to be handled specially just before converting to JS.

Constructors

ElemProps

Fields

epStyle :: Map Text Text

epEvents :: Map Text (ReactEvent -> IO ())

epOtherProps :: Map Text Text

Cannot be: style, key

data ReactElement state

Used only in the ‘DSL’ to create a pure tree of elements.

Constructors

ReactElement

Fields

elemName :: Text

Name.

elemProps :: ElemProps

Properties: style, events, generic.

elemChildren :: Vector (ReactNode state)

Children.

data ReactComponent state

Used in the DSL for constructing an instance of a component class.

Constructors

forall cursor . ReactComponent

Fields

compName :: ReactClass' cursor

compProps :: ElemProps

compLens :: Lens state cursor

data ReactNode state

Also used for the DSL AFAICT, not produced directly, constructed via smart constructors.

Constructors

RNElement (ReactElement state)

RNComponent (ReactComponent state)

RNText Text

toReactElem :: App state m -> ReactNode state -> IO ReactElement'

Convert our DSL ReactElement to a JS element.

genCursor :: TVar Int -> TVar (HashMap Int Cursor) -> Cursor -> IO Int

getPropBool :: JSString -> ReactEvent -> IO Bool

getPropInt :: JSString -> ReactEvent -> IO Int

getPropElement :: JSString -> ReactEvent -> IO Element

getPropEvent :: JSString -> ReactEvent -> IO Event

js_preventDefault :: ReactEvent -> IO ()

js_stopPropagation :: ReactEvent -> IO ()

js_getModifierState :: ReactEvent -> JSString -> IO Bool

newtype EventListener event

Constructors

EventListener Text

class (Coercible ReactEvent event, Coercible event ReactEvent) => IsReactEvent event

type ReactT state m = ReaderT (TVar state) (StateT (ReactNode state) m)

React transformer.

type React state = ReactT state Identity

Pure react monad.

modifyEl :: Monad m => (ReactElement state -> ReactElement state) -> ReactT state m ()

Modify the element in the state.

In the case the current node is text, this is a no-op. Given that the API doesn't have a way to build inside a Text, that's okay. We could GADT the ReactNode type to remove this case, but not worth it ATM. Also there is now a Component type, but I don't know how to build inside one of those either, so it's also a no-op (for now).

modifyProps :: MonadState (ReactNode state) m => (ElemProps -> ElemProps) -> m ()

Modify properties of the element.

runReactT :: Text -> TVar state -> ReactT state m a -> m (a, ReactNode state)

Run the react monad.

data Component state cursor m

A component for some state transforming over some monad.

Constructors

Component (ReactClass' cursor)

data Class state cursor m

Class used for creating components.

Constructors

Class

Fields

classApp :: App state m

Application.

classRender :: ReactT state m ()

Rendering function.

classDidMount :: forall props. Lens state cursor -> JQuery -> JSRef props -> IO ()

Did mount handler.

classDidUpdate :: forall props. Lens state cursor -> JSRef props -> IO ()

Did update.

classShouldUpdate :: forall props. Lens state cursor -> JSRef props -> IO Bool

Should update?

classReceivingProps :: forall props. Lens state cursor -> JSRef props -> IO ()

Receiving new props.